NO_ISSUE: remove unused code and refactor tests structure to have better control on run_if_changed tests triggers on release
