+++ tests/vmaas/test_compute_instance_api_fields.py
@@ -4,9 +4,9 @@ import time
 
 import yaml
 
-from tests.helpers import wait_for_deletion, wait_for_provision, wait_for_running
-from tests.k8s_client import K8sClient
-from tests.runner import poll_until
+from tests.core.helpers import wait_for_deletion, wait_for_provision, wait_for_running
+from tests.core.k8s_client import K8sClient
+from tests.core.runner import poll_until
 
 TEST_IMAGE_REF: str = "quay.io/containerdisks/fedora:latest"
 TEST_CORES: int = 2
@@ -83,16 +83,12 @@ def test_compute_instance_api_fields(k8s_hub_client: K8sClient, k8s_virt_client:
     assert vm_status == "Running", f"VM should be Running, got {vm_status}"
 
     # Immutability: cores
-    output, rc = k8s_hub_client.patch(
-        resource="computeinstance", name=instance_name, patch='{"spec":{"cores":8}}'
-    )
+    output, rc = k8s_hub_client.patch(resource="computeinstance", name=instance_name, patch='{"spec":{"cores":8}}')
     assert rc != 0, "cores field should be immutable"
     assert "cores is immutable" in output, f"Expected immutability error, got: {output}"
 
     # Immutability: memoryGiB
-    output, rc = k8s_hub_client.patch(
-        resource="computeinstance", name=instance_name, patch='{"spec":{"memoryGiB":16}}'
-    )
+    output, rc = k8s_hub_client.patch(resource="computeinstance", name=instance_name, patch='{"spec":{"memoryGiB":16}}')
     assert rc != 0, "memoryGiB field should be immutable"
     assert "memoryGiB is immutable" in output, f"Expected immutability error, got: {output}"
