+++ tests/vmaas/test_compute_instance_restart_negative.py
@@ -3,10 +3,10 @@ from __future__ import annotations
 import time
 from datetime import UTC, datetime
 
-from tests.osac_cli import OsacCLI
-from tests.grpc_client import GRPCClient
-from tests.helpers import wait_for_cr, wait_for_deletion, wait_for_restart, wait_for_running
-from tests.k8s_client import K8sClient
+from tests.core.grpc_client import GRPCClient
+from tests.core.helpers import wait_for_cr, wait_for_deletion, wait_for_restart, wait_for_running
+from tests.core.k8s_client import K8sClient
+from tests.core.osac_cli import OsacCLI
 
 
 def test_compute_instance_restart_past_timestamp_ignored(
