+++ tests/vmaas/test_subnet_lifecycle.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from uuid import uuid4
 
-from tests.grpc_client import GRPCClient
-from tests.helpers import (
+from tests.core.grpc_client import GRPCClient
+from tests.core.helpers import (
     wait_for_subnet_cr,
     wait_for_subnet_deletion,
     wait_for_subnet_ready,
@@ -11,8 +11,8 @@ from tests.helpers import (
     wait_for_virtual_network_deletion,
     wait_for_virtual_network_ready,
 )
-from tests.k8s_client import K8sClient
-from tests.runner import poll_until
+from tests.core.k8s_client import K8sClient
+from tests.core.runner import poll_until
 
 
 def test_subnet_lifecycle(grpc: GRPCClient, k8s_hub_client: K8sClient, network_class: str) -> None:
