+++ tests/vmaas/test_virtual_network_lifecycle.py
@@ -2,10 +2,14 @@ from __future__ import annotations
 
 from uuid import uuid4
 
-from tests.grpc_client import GRPCClient
-from tests.helpers import wait_for_virtual_network_cr, wait_for_virtual_network_deletion, wait_for_virtual_network_ready
-from tests.k8s_client import K8sClient
-from tests.runner import poll_until
+from tests.core.grpc_client import GRPCClient
+from tests.core.helpers import (
+    wait_for_virtual_network_cr,
+    wait_for_virtual_network_deletion,
+    wait_for_virtual_network_ready,
+)
+from tests.core.k8s_client import K8sClient
+from tests.core.runner import poll_until
 
 
 def test_virtual_network_lifecycle(grpc: GRPCClient, k8s_hub_client: K8sClient, network_class: str) -> None:
